Modified the gradient_by_layer.py. This time it firstly trains a network without dropout and then continues training with dropout. We want to see if the behaviour of norms that we observed in earlier experiments is a result of initialization or the dropout technique.
We have also added a function that estimates the number of activation regions generated by a network. Unfortunately, early experiments show that it is unfeasible to use it for CIFAR10 or large networks. This is due to high number of generated activation regions and higher time necessary to run the function for larger datasets. We pplan to perform further experiment on FashionMNIST.