I found two issues with the code. First issue that I reported in the previous commit (with the is_point_inside_polygon) originated from floating point arithmetics. The points can have a maximum precision of 1e-15 after extracting from Shapely. However, when calculating the distance_to_segment we project a point onto a line segment and this together with later calculations can result in something that differs in 1e-16. Hence we had to increase epsilon to 1e-15.
The second issue lied with the find_line_segment_intersection, where it turned out that I have forgot to put the second double condition into bracket causing an error.

Currently, I am having an issue with gradients and values of the activation regions. The errors are very low (smaller than 1e-13), but I will need to make sure if it is ok to increase epsilon. If I increase epsilon I find that the find_intersection is working incorrectly again. I will have to see what is the reason behind this.